Revert "Feature/hpp gateway"
