+++ CHANGELOG.md
@@ -1,15 +1,4 @@
 # Release Notes for Klarna Checkout for Craft Commerce
-
-## 2.0.0 - 2020-05-02
-
-> {warning} Caution! This update makes several major changes to the plugin. Test the update in a local environment before updating production.
-
-> {info} Read UPDATE_v2.md when upgrading from 1.x versions to 2.x versions.
-
-- Refactored the entire plugin code base to be easier to maintain
-- Added Hosted Payment Page gateway (#19)
-- Un-truncated error messages from GuzzleHTTP (#18) 
-
 ## 1.1.6 - 2020-01-31
 
 ### Fixed
