+++ composer.json
@@ -1,7 +1,7 @@
 {
   "name": "elleracompany/commerce-klarna-checkout",
   "description": "Klarna Checkout integration for Craft Commerce 2",
-  "version": "2.0.0-BETA",
+  "version": "1.1.6",
   "type": "craft-plugin",
   "keywords": [
     "klarna",
@@ -27,8 +27,8 @@
   "require": {
     "ext-json": "*",
     "craftcms/cms": "^3.1.0",
-    "craftcms/commerce": "^2.2.0",
-    "guzzlehttp/guzzle": "^6.5.0"
+    "craftcms/commerce": "^2.0.0",
+    "guzzlehttp/guzzle": "~6.3.0"
   },
   "autoload": {
     "psr-4": {
