+++ src/Plugin.php
@@ -4,8 +4,6 @@ namespace ellera\commerce\klarna;
 
 use craft\commerce\services\Gateways;
 use craft\events\RegisterComponentTypesEvent;
-use ellera\commerce\klarna\gateways\Checkout;
-use ellera\commerce\klarna\gateways\Hosted;
 use ellera\commerce\klarna\gateways\KlarnaCheckout;
 use yii\base\Event;
 
@@ -26,8 +24,7 @@ class Plugin extends \craft\base\Plugin
         parent::init();
 
         Event::on(Gateways::class, Gateways::EVENT_REGISTER_GATEWAY_TYPES,  function(RegisterComponentTypesEvent $event) {
-            $event->types[] = Checkout::class;
-            $event->types[] = Hosted::class;
+            $event->types[] = KlarnaCheckout::class;
         });
     }
 }
