--- src/templates/settings/includes/_credentials.twig
@@ -1,71 +1,0 @@
-<h2>API Credentials euope</h2>
-{% import "_includes/forms" as forms %}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_eu_uid'),
-    name: 'api_eu_uid',
-    value: gateway.api_eu_uid,
-    errors: gateway.getFirstError('api_eu_uid')
-}) }}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_eu_password'),
-    name: 'api_eu_password',
-    value: gateway.api_eu_password,
-    errors: gateway.getFirstError('api_eu_password')
-}) }}
-
-{{ forms.textField({
-    first: true,
-    label: gateway.getAttributeLabel('api_eu_test_uid'),
-    name: 'api_eu_test_uid',
-    value: gateway.api_eu_test_uid,
-    errors: gateway.getFirstError('api_eu_test_uid')
-}) }}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_eu_test_password'),
-    name: 'api_eu_test_password',
-    value: gateway.api_eu_test_password,
-    errors: gateway.getFirstError('api_eu_test_password')
-}) }}
-
-<!--
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('api_eu_consent_notice'),
-    name: 'api_eu_consent_notice',
-    on: gateway.api_eu_consent_notice,
-    instructions: 'Only applicable for stores based in Germany and Austria' | t('klarna-checkout'),
-    errors: gateway.getFirstError('api_eu_consent_notice')
-}) }}
-
-<h2>API Credentials United States</h2>
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_us_uid'),
-    name: 'api_us_uid',
-    value: gateway.api_us_uid,
-    errors: gateway.getFirstError('api_us_uid')
-}) }}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_us_password'),
-    name: 'api_us_password',
-    value: gateway.api_us_password,
-    errors: gateway.getFirstError('api_us_password')
-}) }}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_us_test_uid'),
-    name: 'api_us_test_uid',
-    value: gateway.api_us_test_uid,
-    errors: gateway.getFirstError('api_us_test_uid')
-}) }}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('api_us_test_password'),
-    name: 'api_us_test_password',
-    value: gateway.api_us_test_password,
-    errors: gateway.getFirstError('api_us_test_password')
-}) }}
--->
