--- src/templates/settings/klarna-hpp.twig
@@ -1,180 +1,0 @@
-{% import "_includes/forms" as forms %}
-
-{% if siteUrl | slice(0,4) != 'http' %}
-    <div class="pane alert" style="background-color: #f8d7da; color: #721c24; border-color: #f5c6cb;">
-        It does not look like the Base URL is configured properly. This might create issues with the generation of merchant URLs for Klarna.<br>
-        Please update the following setting:
-        <br><br>
-        <code>Settings -> Sites -> sitename -> Base URL</code>
-    </div>
-{% endif %}
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('title'),
-    name: 'title',
-    value: gateway.title,
-    errors: gateway.getFirstError('title')
-}) }}
-
-{{ forms.textAreaField({
-    label: gateway.getAttributeLabel('description'),
-    name: 'description',
-    value: gateway.description,
-    rows: 4,
-    errors: gateway.getFirstError('description')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('send_product_urls'),
-    name: 'send_product_urls',
-    on: gateway.send_product_urls,
-    instructions: 'Send product and product image URLs to Klarna' | t('klarna-checkout'),
-    errors: gateway.getFirstError('send_product_urls')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('log_debug_messages'),
-    name: 'log_debug_messages',
-    on: gateway.log_debug_messages,
-    instructions: 'Log debug messages' | t('klarna-checkout'),
-    errors: gateway.getFirstError('log_debug_messages')
-}) }}
-
-{{ forms.elementSelectField({
-    label: gateway.getAttributeLabel('logo_id'),
-    id: 'logo_id',
-    name: 'logo_id',
-    elements: [gateway.getLogoAsset()],
-    elementType: 'craft\\elements\\Asset',
-    selectionLabel: "Choose"|t('app'),
-    limit: 1,
-    errors: gateway.getFirstError('logo_id')
-}) }}
-
-{{ forms.elementSelectField({
-    label: gateway.getAttributeLabel('background_id'),
-    id: 'background_id',
-    name: 'background_id',
-    elements: [gateway.getBackgroundAsset()],
-    elementType: 'craft\\elements\\Asset',
-    selectionLabel: "Choose"|t('app'),
-    limit: 1,
-    errors: gateway.getFirstError('background_id')
-}) }}
-
-<!--
-<div>
-    {{ forms.checkboxSelectField({
-        label: gateway.getAttributeLabel('methods'),
-        instructions: "Select what payment methods you customer are offered"|t('klarna-checkout'),
-        name: 'methods',
-        options: gateway.getAvailablePaymentMethods(),
-        values: gateway.methods,
-        warning: false,
-        errors: gateway.getFirstError("methods"),
-    }) }}
-</div>
--->
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('test_mode'),
-    name: 'test_mode',
-    on: gateway.test_mode,
-    instructions: 'Enable Test Mode' | t('klarna-checkout'),
-    errors: gateway.getFirstError('test_mode')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('test_mode'),
-    name: 'test_mode',
-    on: gateway.test_mode,
-    instructions: 'Enable Test Mode' | t('klarna-checkout'),
-    errors: gateway.getFirstError('test_mode')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('mandatory_date_of_birth'),
-    name: 'mandatory_date_of_birth',
-    on: gateway.mandatory_date_of_birth,
-    instructions: 'Make customer date of birth mandatory' | t('klarna-checkout'),
-    errors: gateway.getFirstError('mandatory_date_of_birth')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('mandatory_national_identification_number'),
-    name: 'mandatory_national_identification_number',
-    on: gateway.mandatory_national_identification_number,
-    instructions: 'Make national identification number mandatory' | t('klarna-checkout'),
-    errors: gateway.getFirstError('mandatory_national_identification_number')
-}) }}
-
-{{ forms.lightswitchField({
-    label: gateway.getAttributeLabel('api_eu_title_mandatory'),
-    name: 'api_eu_title_mandatory',
-    on: gateway.api_eu_title_mandatory,
-    instructions: 'Make customer title mandatory' | t('klarna-checkout'),
-    errors: gateway.getFirstError('api_eu_title_mandatory')
-}) }}
-
-<h2>Merchant URLs</h2>
-
-{{ forms.textField({
-    label: gateway.getAttributeLabel('terms'),
-    name: 'terms',
-    value: gateway.terms,
-    errors: gateway.getFirstError('terms')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('back'),
-    name: 'back',
-    value: gateway.back,
-    instructions: 'Return page when the customer clicks "Go Back" in HPP' | t('klarna-checkout'),
-    errors: gateway.getFirstError('back')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('cancel'),
-    name: 'cancel',
-    value: gateway.cancel,
-    instructions: 'Return page when the customer cancels the payment' | t('klarna-checkout'),
-    errors: gateway.getFirstError('cancel')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('error'),
-    name: 'error',
-    value: gateway.error,
-    instructions: 'Return page when the payment has an error' | t('klarna-checkout'),
-    errors: gateway.getFirstError('error')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('failure'),
-    name: 'failure',
-    value: gateway.failure,
-    instructions: 'Return page when the payment fails' | t('klarna-checkout'),
-    errors: gateway.getFirstError('failure')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('privacy'),
-    name: 'privacy',
-    value: gateway.privacy,
-    errors: gateway.getFirstError('privacy')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('checkout'),
-    name: 'checkout',
-    value: gateway.checkout,
-    errors: gateway.getFirstError('checkout')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('status_page'),
-    name: 'status_page',
-    value: gateway.status_page,
-    instructions: 'Order status page (?)' | t('klarna-checkout'),
-    errors: gateway.getFirstError('status_page')
-}) }}
-{{ forms.textField({
-    label: gateway.getAttributeLabel('success'),
-    name: 'success',
-    value: gateway.success,
-    instructions: 'Order success page. Passes order number as parameter \'number\'' | t('klarna-checkout'),
-    errors: gateway.getFirstError('success')
-}) }}
-{% include 'commerce-klarna-checkout/settings/includes/_credentials.twig' %}
